Address PR review feedback: zvec portability, Zod validation, auth, UI fixes
- Move @zvec/bindings-darwin-arm64 to optionalDependencies (fixes Linux CI/prod)
- Use declared Zod gutQuerySchema instead of manual validation in gut/route.ts
- Add CRON_SECRET auth to process/route.ts (prevents LLM cost abuse)
- Call onRefresh after deletion in LongTermTab and ProceduresTab
- Move getStatusColor/getTypeColor to shared utils.ts from FreshMemoryTab
- Remove z_reports/ dev artifact and add to .gitignore
