+++ app/(sidebar-layout)/(container)/memory/components/FreshMemoryTab.tsx
@@ -34,6 +34,7 @@ import {
 } from '@/components/ui/table';
 
 import { useMemorySessions } from '../hooks/useMemorySessions';
+import { getStatusColor, getTypeColor } from '../utils';
 
 interface FreshMemoryTabProps {
   onRefresh?: () => void;
@@ -101,37 +102,12 @@ export function FreshMemoryTab({ onRefresh }: FreshMemoryTabProps) {
     }
   };
 
-  const getStatusColor = (status: string) => {
-    switch (status) {
-      case 'active': return 'bg-green-500/10 text-green-600 border-green-500/30';
-      case 'completed': return 'bg-blue-500/10 text-blue-600 border-blue-500/30';
-      case 'abandoned': return 'bg-yellow-500/10 text-yellow-600 border-yellow-500/30';
-      default: return 'bg-gray-500/10 text-gray-600 border-gray-500/30';
-    }
-  };
-
   const getOutcomeIcon = (outcome: string | null) => {
     switch (outcome) {
       case 'success': return <CheckCircle2 className="h-3.5 w-3.5 text-green-500" />;
       case 'failure': return <XCircle className="h-3.5 w-3.5 text-red-500" />;
       default: return <AlertCircle className="h-3.5 w-3.5 text-muted-foreground" />;
     }
-  };
-
-  const getTypeColor = (type: string) => {
-    const colors: Record<string, string> = {
-      tool_call: 'bg-purple-500/10 text-purple-600',
-      tool_result: 'bg-indigo-500/10 text-indigo-600',
-      user_preference: 'bg-pink-500/10 text-pink-600',
-      error_pattern: 'bg-red-500/10 text-red-600',
-      decision: 'bg-amber-500/10 text-amber-600',
-      success_pattern: 'bg-green-500/10 text-green-600',
-      failure_pattern: 'bg-red-500/10 text-red-600',
-      workflow_step: 'bg-blue-500/10 text-blue-600',
-      insight: 'bg-cyan-500/10 text-cyan-600',
-      context_switch: 'bg-orange-500/10 text-orange-600',
-    };
-    return colors[type] || 'bg-gray-500/10 text-gray-600';
   };
 
   if (isLoading) {
