+++ app/(sidebar-layout)/(container)/memory/components/LongTermTab.tsx
@@ -92,6 +92,7 @@ export function LongTermTab({ onRefresh }: LongTermTabProps) {
     try {
       await removeMemory(deleteTarget.uuid);
       setDeleteTarget(null);
+      onRefresh?.();
     } catch {
       // Error handled by hook
     } finally {
