+++ app/api/memory/process/route.ts
@@ -1,3 +1,5 @@
+import { timingSafeEqual } from 'crypto';
+
 import { NextRequest, NextResponse } from 'next/server';
 
 import { classifyBatch } from '@/lib/memory/analytics-agent';
@@ -6,7 +8,20 @@ import { EnhancedRateLimiters } from '@/lib/rate-limiter-redis';
 import { authenticate } from '../../auth';
 
 /**
+ * Timing-safe comparison of secret strings to prevent timing attacks.
+ */
+function verifyCronSecret(provided: string | null): boolean {
+  const expected = process.env.CRON_SECRET;
+  if (!expected || !provided) return false;
+  const a = Buffer.from(provided);
+  const b = Buffer.from(expected);
+  if (a.length !== b.length) return false;
+  return timingSafeEqual(a, b);
+}
+
+/**
  * POST /api/memory/process - Trigger analytics agent batch classification
+ * Requires CRON_SECRET header to prevent abuse (each call triggers LLM invocations).
  */
 export async function POST(request: NextRequest) {
   try {
@@ -20,6 +35,13 @@ export async function POST(request: NextRequest) {
 
     const auth = await authenticate(request);
     if (auth.error) return auth.error;
+
+    if (!verifyCronSecret(request.headers.get('x-cron-secret'))) {
+      return NextResponse.json(
+        { success: false, error: 'Forbidden: this endpoint requires cron authorization' },
+        { status: 403 }
+      );
+    }
 
     const result = await classifyBatch(auth.activeProfile.uuid);
     return NextResponse.json(result);
