+++ package.json
@@ -78,7 +78,6 @@
     "@types/lodash": "^4.17.24",
     "@types/prismjs": "^1.26.6",
     "@types/sanitize-html": "^2.16.0",
-    "@zvec/bindings-darwin-arm64": "^0.2.0",
     "@zvec/zvec": "^0.2.1",
     "bcrypt": "^6.0.0",
     "class-variance-authority": "^0.7.1",
@@ -168,6 +167,9 @@
     "tailwindcss": "^3.4.18",
     "tsx": "^4.21.0",
     "typescript": "^5.9.3"
+  },
+  "optionalDependencies": {
+    "@zvec/bindings-darwin-arm64": "^0.2.0"
   },
   "pnpm": {
     "ignoredBuiltDependencies": [
